# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Foohut Documentation Platform API
  version: 1.0.0
  description: |
    AI-native documentation platform API with hierarchical content organization,
    block-based editing, Git synchronization, and AI-powered features.

    ## Architecture
    - **Organization** > **Collection** > **Space** > **Page** > **Block**
    - JWT-based authentication with role-based access control
    - Cursor-based pagination for list endpoints
    - Rate limiting with standard headers

    ## Authentication
    All endpoints (except auth) require a valid JWT Bearer token.
    Include the token in the Authorization header: `Bearer <token>`
  contact:
    name: Foohut API Support
    email: api@foohut.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.foohut.com/v1
    description: Production server
  - url: https://staging-api.foohut.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /spaces/{spaceId}/pages:
    get:
      tags:
        - Pages
      summary: List pages
      description: List all pages in a space (hierarchical tree structure)
      operationId: listPages
      parameters:
        - $ref: '#/components/parameters/SpaceId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: parentId
          in: query
          description: Filter by parent page ID (null for root pages)
          schema:
            type: string
            format: uuid
            nullable: true
        - name: includeContent
          in: query
          description: Include page content in response
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Pages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Pages
      summary: Create page
      description: Create a new page in a space
      operationId: createPage
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageRequest'
            example:
              title: Getting Started
              slug: getting-started
              parentId: null
              content:
                blocks:
                  - type: heading
                    level: 1
                    content: Getting Started
                  - type: paragraph
                    content: Welcome to our documentation!
      responses:
        '201':
          description: Page created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Page slug already exists in this space
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /pages/{pageId}:
    get:
      tags:
        - Pages
      summary: Get page
      description: Retrieve page details by ID
      operationId: getPage
      parameters:
        - $ref: '#/components/parameters/PageId'
        - name: version
          in: query
          description: Specific version number to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Page retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Pages
      summary: Update page
      description: Update page metadata (title, slug, parent)
      operationId: updatePage
      parameters:
        - $ref: '#/components/parameters/PageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePageRequest'
      responses:
        '200':
          description: Page updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Pages
      summary: Delete page
      description: Permanently delete a page (moves children to parent)
      operationId: deletePage
      parameters:
        - $ref: '#/components/parameters/PageId'
        - name: deleteChildren
          in: query
          description: Also delete all child pages
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Page deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /pages/{pageId}/blocks:
    get:
      tags:
        - Blocks
      summary: Get page blocks
      description: Retrieve all blocks from a page
      operationId: getPageBlocks
      parameters:
        - $ref: '#/components/parameters/PageId'
      responses:
        '200':
          description: Blocks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  blocks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Block'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Blocks
      summary: Update page blocks
      description: Batch update multiple blocks in a page
      operationId: updatePageBlocks
      parameters:
        - $ref: '#/components/parameters/PageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - operations
              properties:
                operations:
                  type: array
                  items:
                    $ref: '#/components/schemas/BlockOperation'
                baseVersion:
                  type: integer
                  description: Base version for conflict detection
      responses:
        '200':
          description: Blocks updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  blocks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Block'
                  version:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Blocks
      summary: Add block to page
      description: Add a new block to a page at a specific position
      operationId: addBlock
      parameters:
        - $ref: '#/components/parameters/PageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBlockRequest'
            example:
              type: paragraph
              content: This is a new paragraph block.
              afterBlockId: block-uuid-123
      responses:
        '201':
          description: Block added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  parameters:
...
    SpaceId:
      name: spaceId
      in: path
      required: true
      description: Space unique identifier
      schema:
        type: string
        format: uuid
    PageId:
      name: pageId
      in: path
      required: true
      description: Page unique identifier
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Pagination cursor for next page
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Validation failed
              details:
                - field: email
                  message: Invalid email format
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer realm="api"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              description: Additional error details
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            requestId:
              type: string
              description: Request ID for support reference
...
    Page:
      type: object
      properties:
        id:
          type: string
          format: uuid
        spaceId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
        description:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        path:
          type: string
          description: Full path from root (e.g., /guides/getting-started)
        status:
          type: string
          enum: [draft, published, archived]
        visibility:
          type: string
          enum: [public, private, unlisted]
        version:
          type: integer
        publishedVersion:
          type: integer
          nullable: true
        authorId:
          type: string
          format: uuid
        lastEditorId:
          type: string
          format: uuid
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        content:
          $ref: '#/components/schemas/PageContent'
    PageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Page'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    CreatePageRequest:
      type: object
      required:
        - title
        - slug
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        emoji:
          type: string
          maxLength: 10
        parentId:
          type: string
          format: uuid
          nullable: true
        content:
          $ref: '#/components/schemas/PageContent'
    UpdatePageRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        emoji:
          type: string
          maxLength: 10
          nullable: true
        parentId:
          type: string
          format: uuid
          nullable: true
...
    Block:
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/BlockType'
        content:
          oneOf:
            - type: string
            - type: object
            - type: array
          description: Block content (structure varies by type)
        properties:
          type: object
          additionalProperties: true
          description: Block-specific properties
        children:
          type: array
          items:
            $ref: '#/components/schemas/Block'
          description: Nested blocks (for list items, toggles, etc.)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...